toolkit/components/alerts/test/test_alerts_noobserve.html

<!DOCTYPE HTML>
<!-- Any copyright is dedicated to the Public Domain.
     http://creativecommons.org/publicdomain/zero/1.0/ -->
<html>
<head>
  <title>Test for Alerts Service</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script> 
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>

<body>
<p id="display"></p>

<br>Alerts service, without observer "asynchronous" case.
<br>
<br>A notification should soon appear somewhere.
<br>If there has been no crash when the notification (later) disappears, assume all is good.

<pre id="test">
<script class="testbody" type="text/javascript">

const Cc = SpecialPowers.Cc;
const Ci = SpecialPowers.Ci;

var gNotificationIsAvailable;

if (!("@mozilla.org/alerts-service;1" in Cc)) {
  todo(false, "Alerts service does not exist in this application");
} else {
  ok(true, "Alerts service exists in this application");

  var notifier;
  try {
    notifier = Cc["@mozilla.org/alerts-service;1"].
               getService(Ci.nsIAlertsService);
    ok(true, "Alerts service is available");
  } catch (ex) {
    todo(false, "Alerts service is not available.", ex);
  }

  if (notifier) {
    try {
      notifier.showAlertNotification(null, "Notification test",
                                     "This notification has no observer");
      ok(true, "showAlertNotification() succeeded");

      gNotificationIsAvailable = true;
    } catch (ex) {
      todo(false, "showAlertNotification() failed.", ex);
    }
  }
}

if (gNotificationIsAvailable) {
  ok(true, "Notification should actually display during later tests... (Hope this doesn't disturb them!)");
}
</script>
</pre>
</body>
</html>
